update migrations, and add pg client

+++ migrations/20241017131254_migrations_auth_.sql
@@ -10,8 +10,8 @@ create table chat
 create table chat_user
 (
     id         serial primary key,
-    id_chat    integer references chat (id) on delete cascade,
-    id_user    integer,
+    id_chat    bigint references chat (id) on delete cascade,
+    id_user    bigint,
     created_at timestamp,
     updated_at timestamp
 );
@@ -19,7 +19,7 @@ create table chat_user
 create table messages
 (
     id         serial primary key,
-    sender     integer,
+    sender     bigint,
     text       text,
     id_chat    integer references chat (id) on delete cascade,
     created_at timestamp,
